+++ chirho/dynamical/internals/backends/torchdiffeq.py
@@ -15,8 +15,7 @@ from chirho.dynamical.internals.solver import (
     simulate_point,
     simulate_trajectory,
 )
-from chirho.dynamical.ops_ import InPlaceDynamics
-from chirho.dynamical.ops_.ops import State, Trajectory
+from chirho.dynamical.ops import InPlaceDynamics, State, Trajectory
 
 S = TypeVar("S")
 T = TypeVar("T")
