+++ chirho/dynamical/ops.py
