--- chirho/dynamical/ops_/__init__.py
@@ -1,1 +1,0 @@
-from .ops import InPlaceDynamics, State, Trajectory, simulate  # noqa: F401
