+++ tests/dynamical/test_solver.py
@@ -6,7 +6,7 @@ import torch
 
 from chirho.dynamical.handlers import InterruptionEventLoop
 from chirho.dynamical.handlers.solver import TorchDiffEq
-from chirho.dynamical.ops_ import State, simulate
+from chirho.dynamical.ops import State, simulate
 
 from .dynamical_fixtures import bayes_sir_model, check_states_match
